Add files via upload
The updated version of the sentiment analyzer adds additional features and sophistication. To begin with, this version removes positive adjectives that are proceeded by a negative modifier. For example, the word "great" will no longer be counted as positive it is proceeded by a "not," so things like "not good", "not great," etc., are screened-out. Secondly, there is a simple graphing function to chart the divergence at points-in-time between share prices and sentiment (or, as positive adjectives % of total document as a proxy).